backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-socketio
python-chess
jsonpatch
aiofiles
pydantic-settings
